Don't force computation of a Variables struct when animations are involved.
Without this change, we can end up asserting in ComputeVariablesData and crashing in CSSVariableResolver::Resolve
due to not finding any variable declarations on the rules we matched, when we have content like in the crashtest
added here, i.e. variables inheriting into a pseudo like ::first-line and animations on the element inside the ::first-line.

We could solve this alternatively by removing the assertion and making CSSVariableResolver::Resolve handle a null
aDeclarations more gracefully, but since we can save the effort of recomputing the Variables struct in this case,
we may as well do it this way.