Prevent the home page from being loaded when we are restoring a session.
The "average case" is clearly not that loading the session file finishes before the load of the initial tab is started (quite the opposite).
Instead of defaulting to false when the session store has not yet been loaded, we make this dependent on the user preference instead which is a definitive flag that the home page in the first tab WILL be overridden.
This resolves #103.