Make the SVG scaling limit for the surface cache more lenient.
Increase the limit to a royal 36MB/element as cap (if the SVG is square). This still avoids performance regressions when scaling large vector images, but also allows for large SVG backgrounds to be cached.
We'll have some performance loss here on e.g. the Chalkboard test (1-2 seconds) because of increased cache management, but it's still considerably faster and smoother than Firefox.
At the same time it allows for the bad practice of slapping a large SVG on a site as html background, considering they are likely, even when using large sizes for "responsive" layout, not going to exceed 3000 px in the largest dimension (and if they do, the web designer needs to be slapped with a big trout).